Refactor business handling into dedicated view models
