+++ IOS/Features/Home/HomeViewModel.swift
@@ -2,13 +2,10 @@ import Foundation
 import CoreLocation
 
 @MainActor
-final class BusinessViewModel: ObservableObject {
+final class HomeViewModel: ObservableObject {
     @Published var topRated: [Restaurant] = []
     /// Results after applying filters and sorting
     @Published var searchResults: [Restaurant] = []
-    @Published var selectedBusiness: Restaurant?
-    @Published var promotions: [Promotion] = []
-    @Published var reviews: [Review] = []
     @Published var errorMessage: String?
     @Published var searchTerm: String = ""
     @Published var selectedFilter: FilterType?
@@ -68,30 +65,6 @@ final class BusinessViewModel: ObservableObject {
                 allResults = try await service.getByTag(tag)
                 applyFiltersAndSort()
             }
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    func fetchBusiness(id: String) async {
-        do {
-            selectedBusiness = try await service.getBusinessById(id)
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    func fetchPromotions(businessId: String) async {
-        do {
-            promotions = try await service.getBusinessPromotions(businessId)
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    func fetchReviews(businessId: String) async {
-        do {
-            reviews = try await service.getBusinessReviews(businessId)
         } catch {
             errorMessage = error.localizedDescription
         }
@@ -251,3 +224,4 @@ enum FilterType: String, CaseIterable {
         }
     }
 }
+
