+++ IOS/Features/RestaurantDetail/RestaurantDetailViewModel.swift
@@ -1,0 +1,43 @@
+import Foundation
+
+@MainActor
+final class RestaurantDetailViewModel: ObservableObject {
+    @Published var selectedBusiness: Restaurant?
+    @Published var promotions: [Promotion] = []
+    @Published var reviews: [Review] = []
+    @Published var errorMessage: String?
+
+    private let service = BusinessService()
+
+    func fetchBusiness(id: String) async {
+        do {
+            selectedBusiness = try await service.getBusinessById(id)
+        } catch {
+            errorMessage = error.localizedDescription
+        }
+    }
+
+    func fetchPromotions(businessId: String) async {
+        do {
+            promotions = try await service.getBusinessPromotions(businessId)
+        } catch {
+            errorMessage = error.localizedDescription
+        }
+    }
+
+    func fetchReviews(businessId: String) async {
+        do {
+            reviews = try await service.getBusinessReviews(businessId)
+        } catch {
+            errorMessage = error.localizedDescription
+        }
+    }
+
+    func markReviewViewed(_ reviewId: String) async {
+        do {
+            _ = try await service.setViewed(reviewId)
+        } catch {
+            errorMessage = error.localizedDescription
+        }
+    }
+}
